backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]
python-multipart==0.0.6
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-barcode==0.14.1
reportlab==4.0.7
fpdf2==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
